Initial commit - Final Sprint Project

+++ src/main/java/com/KeyinDSA/airline/AirlineRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AirlineRepository extends CrudRepository<Airline, Integer> {
+public interface AirlineRepository extends CrudRepository<Airline, Long> {
     Airline findByCode(String code);
     Airline findByName(String name);
     Airline findByCountry(String country);
